backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlmodel==0.0.22
python-multipart==0.0.9
orjson==3.10.7
loguru==0.7.2
numpy==1.26.4
scikit-learn==1.5.2
sentence-transformers==3.0.1
# Optional (recommended for fast dense retrieval). If install fails on your OS,
# the app will fall back to a pure-numpy cosine search.
# faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0
transformers==4.44.2
accelerate==0.34.2
torch>=2.2.0
sse-starlette==2.1.3
regex==2024.7.24
